backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
openai==1.82.0
APScheduler==3.10.4
python-dotenv==1.0.0
pytz==2023.3.post1
email-validator==2.1.0.post1
gunicorn==21.2.0
markdown2==2.4.12 